Ai-Track-Mohith-main/hirelogic_agent/requirements.txt
google-adk>=1.0.0
google-generativeai>=0.8.0
docling
pytesseract
httpx
python-dotenv
asyncpg
sqlalchemy[asyncio]
pydantic>=2.0
loguru
reportlab
